Remove interfaces and revert to concrete types in auth
There's no point in an interface here except boilerplate. I found it important not to do a git revert so I could explain my reasoning.

Concrete types here are smarter than doing the interfaces for the simple reason that testing with an interface makes no sense. At that point what you're testing is if the service works or not. We know that because we can test the service as a standalone already.

The second reason is that the interface is quite testable by passing stub repositories around anyway.