Make tests for the circles service
The following tests are failing:

You should be able to change the display name of a member.
You should be able to remove a member from a circle.
You should be able to add a member to a circle.

these are due to issues with (my understanding of) Skunk's concurrency