Use a new session for each query to allow for concurrency.
This is a large change that impacts nearly every layer in the application because a top level dependency was changed. The general idea is that each query uses it own session. That means we need to push the session pool all the way down to the queries. This also means that each query prepares the statement right before it executes it.

Since queries are no longer prepared beforehand making a repository is no longer an effectful operation.